+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Added "Hide Day" context menu support for the day headers (when more than 1 day is visible).
 - The "useOnlyDotEventsForMainDisplay" setting now overrides the "maximumEventsPerDayDisplay" setting.
 - Removed lots of duplicate code from the core that was no longer required.
+- Added addHolidays() parameter "triggerEvent", which will fire the "onOptionsUpdated" event (if available).
 - 
 
 ### Version 1.3.0: